Fix missing .cc for StubPlayerComponent in build
This file was renamed from stub_player_components_impl.cc to .h to make
it possible to run player_filter_tests against the stub audio/video
decoders.

However, some methods were not appropriately moved, and the file was
also not removed from the gyp configuration.

This change restores the .cc as well as the static methods that were
originally defined there. This fixes failing builds for stub and
linux-x64x11 configs.

b/138311811

Change-Id: Ia77556941c33a26be30b4878605dc9a0e5e171e9

Former-commit-id: f750f6cc10622c2b66040993a654bf980e4f7f3c