Fix AudioDecoderTest.SingleInputHEEAC sbversion-10
Changes to StubAudioDecoder and PlayerComponents to allow the created
audio decoder to know the codec in use. This allows us to check for AAC
and correctly spoof the required number of output frames, for the
purposes of passing tests.

b/138322536

Change-Id: I4cc358062d8e916ae719e27cbebe7254453bc3f5
(cherry picked from commit 1b76d624a8ced5772a4904f2e8533cb86fe37384)

Former-commit-id: 2deef376a4f728321cca52f5ee05143ef7043c53