Revive the debug console for mozjs.
Restore our own implementation of RuntimeAgent to be used when not using
V8, which implements the Runtime domain in its inspector. The
implementation here has been updated to align with changes to the
debugger architecture since it was deleted.

This restored implementation does not include the glue needed for
dom_agent.js to call |CreateRemoteObject| in its implementation of the
|DOM.resolveNode| command, so some functionality in the "Elements" panel
of the remote DevTools may be broken.

The RuntimeAgent restored here was deleted in:
https://lbshell-internal-review.git.corp.google.com/c/cobalt/+/125282/

b/136201377

Change-Id: I1dbf4aea5b53e006249c699b5b626713a53ab7c8

Former-commit-id: 4ac680a2b335ae3274834745a4b720ec247f3831