[raspi0] Enable various player filter tests
The entire suite of player_filter_tests has been disabled because they
were crashing the entire suite because of a rogue test somewhere in the
group. Because it is hard to reproduce, instead we start by disabling
all of them, and reenabling them in blocks until test failures emerge.

In this change, we will reenable all tests except the tests for
VideoDecoder.

b/138579487

Change-Id: Ibb5f5b12e7b7bed51c2d8f2cea0b5202528af66a

Former-commit-id: 7db23d92c4745643470d9ad113bab7d59c5b546d