Return ERR_FAILED when UDP socket watching fails
Following the steps described in the following bug, we can reproduce
a SbSocketWaiterAdd failure. Since this function sometimes does nothing
on failure, the last socket error could be ERR_IO_PENDING which is not
an acceptable net_log entry, but we still want to at least return the
general ERR_FAILED to indicate that the socket watching failed.

b/138397314

Change-Id: Ia8980c8a4346a7efd21ccb2867308e5eb7bf61d8

Former-commit-id: 88182b5c98ff679faeca798058ba5845f156b2d8