[raspi0] Re-enable certain player filter tests
The entire suite of player_filter_tests has been disabled because they
were crashing the entire suite because of a rogue test somewhere in the
group. Because it is hard to reproduce, instead we start by disabling
all of them, and reenabling them in blocks until test failures emerge.

In this change, we will reenable the MediaTimeProviderImpl and
AudioRenderer tests.

b/138579487

Change-Id: I7b2e4d4830ce1d0fa388346acaf7a496b17f59ed

Former-commit-id: 844053120b7a6e82a62bc32fada5f3a521ada75a