Support gcc/g++ as host toolchain.
This changes the linux-x64x11-gcc-6-3 to also use gcc/g++ for the host
toolchain.

This fixes a build flag for v8 mkshapshot that is not supported or needed
with gcc to be applied only when clang is used.

This fixes b/135211231

Change-Id: Ibb87635d01966bb54fc2fffb10499072ba0bcc9f
(cherry picked from commit 63c92111ee38bba06ac7af147362483b5a209f3d)

Former-commit-id: a2453e9633266f701ed32428f1e182dbead378fa