Have intersection observers account for transforms
Currently, an intersection observer event does not get fired if
an intersection change has occurred as the result of a transform.
This change makes sure that intersection calculations rely on
transformed box properties, and adds related layout tests to make
sure that events get fired upon completion of a transform.

b/138684502
b/139158440

Change-Id: I0a6bc08eb7a42f3f65637357def74ed6b71cb3af

Former-commit-id: c0852665943315644c82fa042a7866de7fcb32f2