Have intersection observers account for transforms
Currently, an intersection observer event does not get fired if
an intersection change has occurred as the result of a transform.
This change makes sure that intersection calculations rely on
transformed box properties, and adds related layout tests to make
sure that events get fired upon completion of a transform.

b/138684502
b/139158440

Change-Id: I0a6bc08eb7a42f3f65637357def74ed6b71cb3af

Former-commit-id: 713ee48321c7cc2d108630de0863b964081867d5