Also consider siblings when clearing rule matching state.
Also clear the rule matching state of siblings of the root element.

When there are two sibling <html> elements in the document, still only one can be the root element,
resulting in some rule matching state not being cleared unless it is also cleared for siblings of
the root element.

This fixes b/134768217

Change-Id: I5854d47c26a9e33b7988804a329923ed181b9dd2

Former-commit-id: 7e1a9c1cfbb4e12172764a3987a023887d921020