[raspi0] Re-enable certain player filter tests
The entire suite of player_filter_tests has been disabled because they
were crashing the entire suite because of a rogue test somewhere in the
group. Because it is hard to reproduce, instead we start by disabling
all of them, and reenabling them in blocks until test failures emerge.

In this change, we will reenable the AdaptiveAudioDecoder tests.

b/138579487

Change-Id: Idbd5d40970422045e3f1c54098beb1df21ae2489

Former-commit-id: 73b52e4673f259657f0b9bf9faed76cb9664e0dc