Remove spammy logs from UDPSocketStarboard::SetDoNotFragment
SetDoNotFragment is called on every QUIC stream initialization, so the
NOTIMPLEMENTED inside is really spammy in debug/devel builds. This
function is not necessary for net or QUIC to work properly but may
introduce a very slight performance increase. Since Starboard does
not support implementing this function yet, we will just disable the
log and put a comment in its place.

b/138795624

Change-Id: Ieaed0e5e37333343a552b64929b924dda77fc5bb

Former-commit-id: 89518174d2cfbdbd39ee4d027dc55f1ee828337d