Disable wasm API in V8 globally
Cobalt does not support WebAssembly and the implementation is buggy for
sure but we did not disable the native wasm web API in V8 in the past.

I found two very very convenient V8 flags that are designed to disable
wasm API.

With this CL, the Clusterfuzz crash below is not reproducible anymore.

b/134768726

Change-Id: I0115ef1d00fdcfd447f74505f4d9b949ad08ec1a

Former-commit-id: a35bfc999d1ef03acb451c09ce197287659103bd