Fix toolchain compilation for linux-x64x11-gcc-6-3.
GCC version 6.3 fails to compile with GCC 7.2 that is currently installed on
our workstations.

This adds patches to the GCC 6.3 source file to fix the compilation errors.
The patches include a type change for 'struct ucontext', and disabling of
the libsanitizer support to avoid the large number of incompatibilities that
it has.

This also adds some checksums to verify the downloaded files, and silences
some harmless compiler warnings.

This fixes b/139374019

Change-Id: I6509f16adce2cd99f5f840f7060339e10b1ac1b7

Former-commit-id: 088928fe5f779500ecec82e387d4fefe3c02d6e1