Add 'memory_tracker' debug console command.
This adds a 'memory_tracker' debug console command that can be used to start
a memory tracker.

This adds the option to specify an output filename for the memory_tracker
log output, by appending ':filename' to the tracker identifier.

On Xbox One, change the kSbSystemPathDebugOutputDirectory to use the
kSbSystemPathCacheDirectory instead of the kSbSystemPathTempDirectory,
because the temp directory is wiped as soon as the application exits.

b/139489834

Change-Id: I1b71a1ef4597afd513b2adb601c352a51859974d

Former-commit-id: 412569069a14a33660618e28ff79b67e49621813