Remove SbAudioSinkCreate usage in AudioRenderer
Use the SbAudioSinkPrivate::Type counterpart methods directly in
AudioSinkRendererImpl which is owned by the filter based SbPlayer.

This allows us to decouple the WebAudio needs from SbPlayer's for the
SbAudioSink interface.

AudioRenderer does not need to depend on the SbAudioSink at all, but
this necessitates adding a new class that can be used both by
SbAudioSinkPrivate and AudioRenderer.

b/137395241

Change-Id: I173b759b090dca9e7879d53a36083d3f96fab12f

Former-commit-id: ae82f32bc9b9970450b26cc1961ee984db03750d