Flexbox: Column flex container sizes.
This fixes sizing of the flex container in edge cases with indefinite
sizes, both in cross and main axis.

This fixes sizing of the container main axis to fit-content for column
flex containers when there is no definite main size (including when
there is a definite max main size).

This adds corresponding unit tests in a combined page.

This fixes b/137045624

Change-Id: I530a7f2da5b2b516919374bf7fa574dc1d6c4944

Former-commit-id: 5d743565f73269a98758d2588a5ce9871c78b163