Fix libc & libm incompatabilities on Clusterfuzz
Use a ".symver" assembler directive to bind quick_exit() to the older
implementation that was found in Ubuntu 16 xenial's libc.so v2.23 (and
back to v2.10) rather than requiring the updated implementation that's
now available in libc.so v2.28 (and back to v2.24) found on our build
machines running gLinux.

https://sourceware.org/git/gitweb.cgi?p=glibc.git;h=47dd3543d36465496970406da03db5aecdc377ee

We still need to copy the newer libm.so into the local lib/ directory to
be used on Clusterfuzz since there are a few functions that have been
updated since the version found on Ubuntu 16 xenial.  The math functions
are so ubiquitous that the same trick of binding to the older versions
doesn't work. Even using the compiler's --include flag to force a header
with the ".symver" directives into every compilation unit doesn't catch
them all, resulting in a loader version mismatch later in Cobalt's
startup rather than right at the beginning.

No other libraries have compatibility problems, so they don't need to be
copied into the local lib/ directory.

Note: The dynamic symbols used by the binary, including the symbol
versions, can be seen by running "objdump -T" on it.

b/139670583

Change-Id: I0fdcfa4f32bb1e006ab7fbc80a76b2e54d2ff598

Former-commit-id: abbe9ee5e1e6aebba8600b02e9e3d7847d7bc082