Use WeakPtr to protect MicrophoneAudioSource functions
This is a followup change for this CL:
https://lbshell-internal-review.git.corp.google.com/c/cobalt_src/+/143560

We can not PostTask of MicrophoneAudioSource with scoped_refptr from
another thread since the class is not thread-safe.
We also can not create a scoped_refptr on MainWebModule and use it with
PostTask from MicrophoneManager because MicrophoneManager is owned by
MicrophoneAudioSource and storing a scoped_refptr of
MicrophoneAudioSource in MicrophoneManager creates reference cycle.
Using WeakPtr is different from the logic before the last CL, but it
should be safe because the affected functions essentially just inform
the consumer of critical events. When the consumer destroyes
MicrophoneAudioSource it is already meaningless to invoke the stored
callbacks in MicrophoneAudioSource.

b/140457640

Change-Id: I6364f83ccd9eb067c905fc07aee1397675e9800f

Former-commit-id: 5b487177f391e6924fbd9a1785d00ca8fc823459