Fix a Wxor-used-as-pow false positive
Since people seem to write "2 ^ X" hoping that it means "1 << X", clang
recently added a warning for this pattern.

It incorrectly fires on this file. To suppress it, restructure the code
to be less clever. (Alternatively we could use "xor" instead of "^" or
write "0x2" instead of "2" but both seem worse.)

No intended behavior change.

This is a patch from:
https://chromium-review.googlesource.com/c/webm/libwebp/+/1786630

b/140666417

Change-Id: I9e4c264433f880bacfde1887313090810de15064
(cherry picked from commit 128f689d966728411b349e1a8037eeee36b5055d)

Former-commit-id: 082ddbb8a2bd260f50b2b849fb9805a0771969c9