Fix intersection observer crash on unobserve()
Check that an element's intersection observer module exists
before attempting to access its list of intersection observers.
Also give a warning instead of an error if the particular
intersection observer is not found in that list, because this
can happen from JavaScript if a particular intersection observer
does not call "observe" on an element before attempting to
"unobserve" it.

b/140873337

Change-Id: I5ade74546e7c63cec3dc2b757740426ff2299c7a

Former-commit-id: ae4f3c9b45562dbd6f6e7800c6caee65f067cf82