Rewrite the java logging to use reflection.
The native calls crash with UnsatisfiedLinkError whenever
a component is launched that is not the main activity and
tries to log a message. This can happen for a service or
broadcast receiver for example.

Using reflection instead solves the problem.

b/140820301

Change-Id: Icf279f81999e853207502bea1b5f121f3ad324bd
(cherry picked from commit 6feef229bd756fa3495353911bd849de27cf8f10)

Former-commit-id: 5e041ba4e29eab6957af971d06f27becf7fb2fb1