Fix PS4 crash on intersection observer unobserve()
"unobserving-elements-without-calling-observe-should-not-crash" is
crashing on PS4. When an element gets garbage collected, it releases
the ref counted intersection observer objects stored in
ElementIntersectionObserverModule's intersection observer vectors,
but the destructor for the intersection observer object calls into
ElementIntersectionObserverModule::UnregisterIntersectionObserverForRoot,
which attempts to release the same intersection observer object again.

This change makes sure that the intersection observer objects in the
vectors are destroyed in the dtor for ElementIntersectionObserverModule
so they cannot be attempted to be released again later.

b/140873337

Change-Id: I3df140e1316d08c35a802b4e94a7ec5d8a40783b

Former-commit-id: 571199478fafe9d450c2c6e3005b192a5a0c0d68