Increase allocation increment for Media Buffers
Previously we had an allocation increment of 1MB, but we may try raising
it to 4MB to reduce the overall number of calls to allocate. As a
result, we ensure that the Allocator class is able to handle the
scenario when the allocation increment is greater than the remaining
capacity available to allocate.

b/140813291

Change-Id: Ic41ef76fa4edb6a583cd2e90496c6dc2badc4b34

Former-commit-id: b813086e573dd1fb2722c3fed462256bf1ccdf77