Allow DISABLE_TESTING in either application or platform test filters.
The list of platform/application filters is a mostly-homogeneous list of
test_filter.TestFilter objects, each with a target_name property. However,
since test_filter.DISABLE_TESTING is a string used as a signal not to run any
tests at all, it doesn't have that property. Therefore, don't consider it when
verifying that the filter targets are configured in the correct application vs.
platform config.

b/138579487

Change-Id: I02b067595324c968388746ffb07313e059894547

Former-commit-id: e77a0ddc3c21048938c098d049c6b3e99a58c087