Remove the speech stub sources for evergreen
When building updater_sandbox on evergreen platforms, I got error
"If speech synthesis not enabled on this platform, please exclude it
from the build". Currently evergreen platforms don't support speech
synthesis or speech recognition. I added a condition check to exclude
the speech source files if speech is not supported.

Change-Id: Ie0564a328c6ec170f65547399a1ec58f2eff5b5e

Former-commit-id: cf1b31f448c87506075bf3fe0c86ccd0d5b1f4ce