Remove the speech stub sources for evergreen
When building updater_sandbox on evergreen platforms, I got error
"If speech synthesis not enabled on this platform, please exclude it
from the build". Currently evergreen platforms don't support speech
synthesis or speech recognition. I added a condition check to exclude
the speech source files if speech is not supported.

Change-Id: Ie0564a328c6ec170f65547399a1ec58f2eff5b5e

Former-commit-id: d04b7a60ea525d48aac0a8c2b4a85ed9032df597