Add 'app_launcher_zip' target to package the app launcher.
Update app_launcher_packager.py so that it can emit a list of all
dependencies, and can create a zip rather than just copying to a
directory. This makes it usable from gyp to make a zip file in the out
directory containing the files that are needed to run tests.

This will replace the nightly upload of the app launcher package to X20
so that the runner scripts used by Mobile Harness will be coherent with
the actual code in the branch being tested.

b/141478649

Change-Id: Ia97b059bf9f8caa3746c9e9d9b2f454569f65420

Former-commit-id: 9622d2dc7126122a48d9c4d4dc33845e6ca1b868