[DRM] Handle keybox failures without crashing
Previously, when keyboxes would change (typically for development use
cases or in the event of an emergency), then the app would crash if the
keybox and the provisioning file did not match.

This change address this by computing a hash for the keybox, and storing
it with the provisioning file. Whenever the DRM system is created, we
check that the provisioning and the keybox checksums match. If they do
not, then we clear the provisioning information (forcing re-provisioning
later). Otherwise, things proceed as normal.

b/140836913

Change-Id: If7ce2b0594106f4b7621ce2ab0276cf45caad9a2

Former-commit-id: ca7be72ffcc95e152cd9d44de8a01b90ba5beb25