Enable STL debug mode for most linux builds.
This enables libstdc++ and libc++ debug mode for most linux builds.

Also, add the size & speed optimization flags for the clang-3.6 and gcc-6.3
gold builds.

Note: It is not enabled for the clang-3.6 build, because that version of
clang does not support debug mode well.

This also includes a fix found by this in RectAllocator: Since upper_bound
returns the first element where the comparator is false and that is not
equal, it should suffice to only compare the area in FirstRectIsBigger.

This fixes b/142828237

Change-Id: Ida8b86638cde60c0782cee4d48f0b6a8154ea2e7

Former-commit-id: 3a5435cd32e95f696433d7601db822412c79a65f