Adopt V8 7.7 into Cobalt
Necessary changes in Cobalt to adapt to new V8 7.7 APIs.
Most of the gyp files are based on nodejs's gyp files of the same V8
version. Nodejs uses different gyp runner and have different directory
structure but we are using the same compiled file list and mostly the
same build steps.

WASM module is not removed in this CL yet.

b/130738459


Former-commit-id: 30af83a99add926a6a6a3d66f881f73f9c41d3ba