Refine format selector in web_media_player_sandbox
Instead of reading an arbitrary header and performing a string-search
for a keyword to define the input format, we instead attempt to use the
ChunkDemuxer to operate on the input data while permuting the possible
formats until it succeeds the AppendData() operation.

This makes the code easier to maintain, as just a list of formats
applicable in the sandbox.

b/135536560

Change-Id: I60eeee311ab165bf89527eea1735d4c0ce59f09d

Former-commit-id: 1e14bae63b5737dc6e22a0db454c9d422db56b16