Refine format selector in web_media_player_sandbox
Instead of reading an arbitrary header and performing a string-search
for a keyword to define the input format, we instead attempt to use the
ChunkDemuxer to operate on the input data while permuting the possible
formats until it succeeds the AppendData() operation.

This makes the code easier to maintain, as just a list of formats
applicable in the sandbox.

b/135536560

Change-Id: I60eeee311ab165bf89527eea1735d4c0ce59f09d
(cherry picked from commit 0e334e4c5d1c0d6c98e4b4c709cea3b02dcd720f)

Former-commit-id: 97fc63eb2aad8f571e50db0d60707592a615d5d9