Enable v8 embedded builtins
Turns on v8's option to embedded builtin functions in v8 binary. This
will become mandatory some day by v8 in the future. With embedded
builtins, the builtin functions are no longer stored in snapshots and
different isolates do not need to each keep a copy of all the builtin
functions.

When I was debugging a JIT-less mode startup crash, I realized that
JIT-less v8 has the assumption of using embedded builtins, so this is
also blocking the JIT-less mode.

b/142138175
b/130738459

Change-Id: Ic8837a3bc524125756e384628be6da0a26d89c23

Former-commit-id: 4ac519dedb57d8cc7def3f78710cc886ca84f73a