Remove DCHECK on video codec in DecoderBufferAllocator
It is possible for video codec to be unknown when
DecoderBufferAllocator::Allocate() is called, so handle this case
elegantly.

Also revert the change in ShellDemuxer to call UpdateVideoConfig() when
video decoder config is received, as it is no longer necessary.

Note that we should consider to allow video codec to be
kSbMediaVideoCodecNone for SbMedia buffer functions, by specifying
clearly in the comment.

b/139737138

Change-Id: Ifea73f6cd5f2833ae2fb0a13fd0c6c7773ba6387

Former-commit-id: 0d82059c9cb4cc7fbbb599c5e4b935919e058570