Calculate and output the time to unzip cobalt.crx
This measurement is not perfectly precise. The precise method is to
record the finish time in the callback function of zip::Unzip, which is
ComponentUnpacker::EndUnzipping. But to do that, more changes to the
update_client and zip library are required, which increase the
complexity to rebase in the future. I measured the offset of the current
method: it's only about 0.1ms.

b/144180696

Change-Id: I42182c8db2d4aed4683743f2581446996d40e6db

Former-commit-id: 56808760635a14aaafda8b582b1559d381cc29d7