WebDriver & web debugger listen to IPADDR_ANY by default.
Remove server-specific logic from WebDriver and the remote web debugger
(DevTools) for determining the IP address of the interface to listen to.
Instead, pass it in from the Browser application, and let it be
specified on the command line with a new "--dev_servers_listen_ip"
switch that affects both for WebDriver & the web debugger. (The previous
"--webdriver_listen_ip" switch is deprecated, but still works.)

On Linux make the servers listen only to localhost by default by
injecting "--dev_servers_listen_ip=::1" if it is not already present on
the command line.

Both servers now use a common |HttpServer::GetLocalInterfaceAddress()|
to determine the actual IP address of the interface they're listening to
in order to report it as a URL in CVals. If listening to IPADDR_ANY, it
picks one with a preference to report an IPv4 address.

The custom WebDriver "/session/startscreencast" command now listens to
the same interface as the WebDriver from which it was started, rather
than always listening to IPADDR_ANY.

b/139097909

Change-Id: I5753f7cba61d1e6a1d34742986f43f0938cc6305

Former-commit-id: b2c95d8064071cbc5bb6234fb1edf130e3715d07