Enable test_runner.py for Evergreen platforms
This change enables test_runner.py to run tests and deploy arbitrary
targets in Evergreen mode. This is accomplished by creating an
implementation of the abstract launcher for Evergreen.

In addition to the primary change, paths provided to the ELF Loader are
now considered relative to the ELF Loader's content.

b/143705215

Change-Id: Id72a92b0a54483a62b501ca5f37965c9aa8b485e

Former-commit-id: 1501dcdf8d8d863dfbd683996c243ad1801fd4f5