Refactor boilerplate out of WebDebuggerTest
Instantiate the WebServer,CobaltRunner, and DebuggerConnection using a
new setUpWith() function that enters them as Context Managers and adds
a cleanup task to exit them at the end of the test.

Check debugger breakpoints with a shared assert_paused() method that
waits for the paused event then asserts that the synchronous and
asynchronous call stacks are as expected.

b/123092171

Change-Id: If607f5c4e5abb921e0e63aa46ad394a65b7fe93d

Former-commit-id: aa9e8520c00886b1950506bdc4cb924911f2d7a7