Implement Media Console
Implement a basic console for media dev and debug purposes. When the
debug console is showing, hit CTRL+M to bring up the media console. This
console hides the other consoles and responds to keyboard inputs to
provide detailed status information regarding the player, and controls
to affect the player state. Hit CTRL+O to leave the media console.

b/144186375

Change-Id: I32e71eff9bcab4ce2563a9e712f6881246188aaa

Former-commit-id: 7d336285069284e6b7594bbe98fe8eff1faa51a5