Define ownership of debug::Command objects by using unique_ptr.
Eliminate references to the temporary Command object created in
DebugClient, and don't allow copies of Command objects.

The callback held in the Command holds a reference to the DebugHub when
commands are sent from the console. When copies of the Command object
are destructed, it interacts unfavorably with the ref-counted DebugHub.

b/144801882

Change-Id: I807f6a3d6f58522bc5d0fb42eb8f9c09bc7aa0fd

Former-commit-id: 9a01eb913bd3c58afd160b7d9027454067fdf35b