Highlight nodes in Cobalt when hovering in the DevTools Elements panel.
The DevTools "highlightNode" command moved from the "DOM" to the
"Overlay" domain, so there's a new OverlayAgent to handle it.

Unlike Chrome, we do show negative margins and allow them to overlap
with the border, padding and/or content highlights.

This implementation does not consider CSS transformations, other than
the fact that getBoundingClientRect() returns a rectangle around the
transformed element.

Also implement the "highlightRect" command.

b/145236117

Change-Id: Ia7efb88d3d640e1d297c21832672f03ecf953752

Former-commit-id: e7fa3b50f57cc75fac079146d44859e2076a9dd8