Refactor debug console into classes.
ConsoleManager toggles between the interactive and media consoles.
DebugConsole is the console to enter JavaScript and see logs.
The HUD is a part of DebugConsole, but ConsoleManager helps to show it.

Change visibility of the HTML frames by setting classes on the <body>.
The console setVisible() methods are informative rather than imperative.

b/145745690

Change-Id: I68b1c9457bb5077c275ad54e4bc5deb10ed483ea

Former-commit-id: d896c6056b25da3615e50ff6b818658c49b79d62