Evergreen black box tests, use LauncherParams
This change implements the ability to run black box tests in Evergreen
mode with the newly created LauncherParams. This new packaging of
parameters will make it easier to adopt additional tests moving
forward and will make it trivial to add or remove parameters needed by
the abstract launcher.

b/146009956
b/145519990

Change-Id: I6d1157e7834a63e5f3fa21622f58b9acecba14ae

Former-commit-id: 6e1ed1485a393ceafbb94879dd1356b510482cfe