Fix timing-related crash when focusing with UI navigation
The old approach temporarily set the ui_nav_item_ to null when
the UI engine initiated a focus. However, this could have resulted
in the HTMLElement creating a new ui_nav_item_ when running the
focusing step. If a layout occurred with this new object, it would
be enabled just before the focusing steps finish and reset the
ui_nav_item_ pointer back to the old value.

This would have resulted in a nav item that is active and could
receive callbacks. However, since the HTMLElement no longer knows
about it, the HTMLElement might be destroyed while the orphaned
nav item processed callbacks for the HTMLElement.

b/146084484

Change-Id: Ifa889b3940e86fb6941f49af84fc3454e18b6b96

Former-commit-id: eec9981bbbf2a9d7b525a211f76be8a7dcd25454