Show DOM mutations in the DevTools Elements tree
Set a MutationObserver on all nodes when they are reported to the
frontend. When the child-list, attributes, or character-data is
mutated, send the corresponding events to the frontend to update its
Elements tree.

Change the node store so ID lookups return an an object holding metadata
along with the node itself. Use that to track whether we've already
reported the children of a node in order to decide which event to send
on a "childList" mutation. If we haven't yet reported a node's children
then just report a change in its child count, otherwise report the child
nodes that were inserted/removed.

b/145235971

Change-Id: I8ff1e04cbcb5f156918c1ccdccc65f09ba53e6d2

Former-commit-id: 7c43f03b354348c6e746f817d49ca74ac90ea0d5