Make socket code prefer connections to IPv4 addresses
Some networks may not be properly configured for IPv6, so
preferentially connect to an IPv4 address first (if available).

Also disabled a couple unit tests whose assumptions are no
longer valid with this customization.

b/110164203

(This is a re-submit of go/cobalt-cl/105651 after it was lost in rebase)
(Previous Change-Id:I4121cfdc8ee73a9697e7ec7800fb993f7b921c25)

Change-Id: I489a5e219334723f938052a64ba3d69f7cf9cd27

Former-commit-id: 7af69a1335f903d1c4afd61cb449d1cc7e1da2e6