Update web_debugger black box test to expect events in order received.
wait_event() raises a DebuggerEventError if the next event received is
not the one it's waiting for. This allows us to ensure that events are
received in the order expected (e.g. DOM.childNodeRemoved should be
before DOM.childNodeInserted when changing textContent).

However, it ignores "Debugger.scriptParsed" events unless that's what it
is specifically waiting for since they get sent as a side effect of the
backend implementation detail that it injects its own code.

Added tearDown() to check that there are no unprocessed events when the
test ends. This is also an opportunity to consolidate the common code
to end the test and check the JavaScript test status.

assert_paused() now waits for the expected "Debugger.resumed" event from
it resuming execution before returning. Also put the expectation of the
anonymous stack frame into assert_paused() instead of the boilerplate
every place it's called.

Updated timeout handling for both commands and events to return our own
exception types to show a more meaningful stack trace.

b/145235971

Change-Id: Ibe21f309d915e4da76de0dbcb3dfbafddeb1ce46

Former-commit-id: 1f1b37457a5955d4ac7790bd6b7757447944a5f1