Don't update matching rules on a single element.
For the "CSS.getMatchedStylesForNode" command implementation, don't
update the matching rules on just a single element. That doesn't work
since there's nothing to ensure that the parent's matching rules are
valid. Instead, update the computed style on the element and its
ancestors in the same way that window.getComputedStyle() does.

b/146353490

Change-Id: Ib14e072ce42bd51cc1b3fb25e2e44a4bc4e540c6

Former-commit-id: 87305020a9abc1c1941dc08346c8758b516d6328