Cobalt evergreen version control
Use a version file to record the Cobalt Evergreen version. Bump the
version in this file for each release. Write the update package version
to persisted data after an update package is successfully unpacked. Use
this version from the persisted data in the update check request if it's
not empty and larger than the version of the running binary.

Change-Id: I191e12451e8601c58c9a9104ac3da6dcf909447a

Former-commit-id: d6e7979ea092c3813821232f15777af97fef6481