Add script to create uploadable test data from DMP
In |player_filter_tests| we use DMP files as the basis of running unit
tests against real data. DMP files are stored in two parts:
- As a placeholder, which contains the sha1 of the original DMP.
- As the original DMP, hosted in static storage, under the name alias of
  its sha1.
During the build steps, the DMP file is generated from the static
storage using the placeholder, and is then used in the tests.

The following script makes it easy to produce uploadable and formatted
file versions of the DMP file, that can be commited to the repo and
uploaded to static storage.

b/145691403

Change-Id: I3a33dd77fbfada012777ec038676da0cc3669dcd

Former-commit-id: 3ca11cfdd6351b3a25c6400b91dc68a794729031