Cleanup dom_agent.js prior to reporting DOM mutations.
Updated the node store for more straightforward lookups. It's now a
bi-directional Map that can go to/from DOM nodes and node IDs to avoid
O(n) lookups by iterating an Array. Add a |_getNodeId()| helper to
encapsulate how the Map is used when starting with a DOM node, and use
1-based node IDs so falsey values can indicate not-found.

The node store is maintained as an accurate representation of exactly
the nodes that have been reported to the client. Added a new
|_countChildNodes()| function to get the count of child nodes without
adding them to the node store.

Added a |_reportNodes()| helper to send "DOM.setChildNodes" for both the
"DOM.requestNode" and "DOM.requestChildNodes" commands. Also added a a
recursive |_reportPathFromRoot()| helper for the former command to
ensure all ancestors are reported for the requested node.

Added a special case to always report single text node children even if
they're 1 beyond the depth requested. This takes precedence over the
usual logic of not reporting whitespace text nodes (e.g. the whitespace
between tags in the HTML) that would clutter the DevTools Elements tree.
This meant putting all depth recursion logic into |_getChildNodes()|
rather than part of it being in |_getNodeWithChildren()|.

Renamed |_nodeIsIgnorable()| to |_isWhitespace()| to be more clear what
it's doing, and to mimic Blink's similar implementation.

Updated comments to differentiate actual DOM nodes from the DevTools
Node data object that is sent to the frontend.

Updated the blackbox test by splitting |test_dom()| into two, one for
requesting child nodes from the tree that's been reported, and another
for requesting nodes that haven't been reported through their remote
object id. Fleshed out |test_dom_tree()| to more thoroughly check the
reported children and the childNodeCount, and check for the unrequested
lone text nodes that are special-cased.

JS style changes to use |let| instead of |var|, and simplified function
definitions rather than assigning to variables.

b/145235971

Change-Id: I225447f1b0cb06f0a19a3347bff7f3995f194249

Former-commit-id: 19dbec9ebd43d7512ad5839649e5c96a48b4dbe5