Add new function SbPlayerGetPreferredOutputMode()
It takes both a fully populated |SbMediaAudioSampleInfo|, a fully
populated |SbMediaVideoSampleInfo|, an SbDrmSystem, and the max video
capability associated with the video as parameters, and returns the
preferred output mode for such a video.  This allows the implementation
to fine tune its output mode supported provided to Cobalt, when it
supports both.  For example, an implementation may decide to use decode
to texture mode for 10x videos, or only use punch out mode for hdr
videos.

b/145019717

Change-Id: I67f7eccca8810b91a64f78eca31489cbf5d016ee

Former-commit-id: 8a122ae76f9b83e6f71d05e948880e51fdea3016