Ensure Video DMP files are read in 1MB chunks
Previously the entire file was buffered into memory, and this would
cause low-end devices like the raspi-0 to be drastically slow and crash.
This was changed to read directly from the file instead of buffering at
all, however this negatively impacted nxswitch.

This CL is a compromise where we break up the file reading into 1MB
chunks.

b/145839191

Change-Id: I31809b64887510337c43d7342ba4710ce363cd67

Former-commit-id: 184cbe56900a77474937568ffa1651c6f08422b4