Implement LOG_ONCE and fix NOTIMPLEMENTED_LOG_ONCE
This implements a LOG_ONCE and uses it to reduce some logspam.

LOG_ONCE will log the message only the first time when the statement is
executed.  This also adds the corresponding DLOG_ONCE.

The same method is used to fix NOTIMPLEMENTED_LOG_ONCE to be a single
statement and not eat stream parameters.

b/138843162

Change-Id: Ib0c4d7218f6993a9792284a16975806006eeea21

Former-commit-id: da0fd8ac934bb16572d7861ce4d7a27239158949