Update the Android Gradle plugin to 3.5.3
Override the 'android.injected.build.abi' property to build just the
first ABI supported by the connected device. Otherwise, running from the
IDE will try to build the "armeabi-v7a" for all devices since it's a
supported fallback ABI on devices with a different primary ABI. We don't
want to do that since it means gyp-ing two different Cobalt configs,
keeping them in sync, and waiting for two ninja builds each time a
change is made, which is all just a pain for developers.

Set cmake.buildStagingDirectory whether in a ninja deploy build or
building in the IDE so that the CMake files go in a single directory
that we can cleanup rather than creating the default .cxx/ directory.

b/147105794

Change-Id: Iab3f4258e3a87f3964f7049d2407d68a416977d9

Former-commit-id: 4ee7ed472af0f6062c284b9ac99694a02347c554