Raspi should listen to IFADDR_ANY since it's not a Desktop PC
WebDriver and DevTools listen to IFADDR_ANY by default on all platforms
that do not classify themselves as a Desktop PC (because Desktop PCs
are typically developer workstations that are accessed locally through
connections to "localhost" and shouldn't expose themselves to potential
port-scans). Raspi is used as a remote target, and should listen to
IFADDR_ANY by default so that workstations can connect to them on their
actual IP address.

b/147109836

Change-Id: I69d214770ed64a0e642c7ad7bb8a77076159baab

Former-commit-id: 69f981c7b6c32d58c3705db6dd2d6a498bd3c89b