Revert "Revert "Determine map to mesh support at runtime""
This reverts commit 9c657e65de2ebe58b4face1cbfb5b20f7c88f9a9.

Reason for revert: <Fixed black box tests b/147505360>

Though this is a revert, it changes the design of this cobalt
extension significantly so that there is only one point of truth for
whether or not map to mesh is enabled.
This change also refactors pixel_tests to not rely on the macro.

Change-Id: I29c17d9c904d4f8b25247676f85beef7c7237474

Former-commit-id: a11057e3f9dc054caab0178ec2edda82cc456d95