Add woff2 Roboto fonts and font load options.
You can run Cobalt with command flag
--font-format=[woff2|ttf|woff2-preferred|ttf-preferred].
This will be used by ***REMOVED*** in order to gradually switch over to
woff2 fonts, as well as to monitor performance impact of loading
woff2.

This change also sets the Freetype FT_CONFIG_USE_BROTLI option because
the option needs to be defined in order for Freetype to fully
decompress WOFF2 streams.

b/138597763

Change-Id: I1ab437507a6349d619d6ec61a75560becbe8924f

Former-commit-id: a5b7f2ab53c9eb18b2cc7d7618ae40f16dc50d4e