Implement video input frame hash dumping
It is controlled by command line parameter "--dump_video_input_hash" in
non-gold build.  When it is enabled, Cobalt will dump the timestamp and
hash of every audio and video input to the log.  The dumped information
can be later used to check if the inputs are identical across multiple
runs, or check if the decrypted result is the same as the original file,
if both the encrypted file and the original file are played.

b/142016911

Change-Id: Id1838d511a3935b5c09913212872d5bb0d495a88

Former-commit-id: 2d82f51ec15354240d4f1fc81621d7af6cf9d5e8