Create --loader_out_directory, use in test_runner.py
The --loader_out_directory will be identical to the --out_directory,
except that it will be used when building and running in Evergreen mode.
The value this flag normally provides is typically generated using
Starboard tooling, however, when running tests on mobile harness we do
not know where the out directory for the loader will be located. This
flag allows us to specify this location.

b/145519990

Change-Id: I5a9f645693b11eb95228f19397e2218a4c9139cf

Former-commit-id: a32a5cc25410855378257086f641a2e9ba701c8c