Increase robustness of GraphicsSystemTest.
The GraphicsSystemTest sometimes fails if the initialization is fast but
very variable.

The purpose of the test is to catch initialization that gets very slow.
Increasing the measurement time gets a more stable reference time and adding
a floor will reduce flakyness form systems where the initialization time is
fast but relatively variable.

b/147314816

Change-Id: I51a63e5cf5693ef789515238815b4561ea166f7e

Former-commit-id: c834b9d6f0a7ead6ee2bc30f410af56393120825