[Skia m79] Use correct texture format
We need to provide a sized internal format that specifies the number
of color components in a texture, but we are currently providing a base
internal format instead. The incorrect format type ultimately resolves to
GrGLFormat::kUnknown, which leads to the Skia backend not being able to
render images.

The changes here convert the base internal format to the sized internal
format so that the format of the backend texture is properly initialized,
and renames an existing function for greater clarity. This fixes the
image-related pixel test failures.

Change-Id: I4eb501153c6527be074139f7e54b40c16e46077a

Former-commit-id: 18e4ffc70220f8bf901b9ac3eff3b65cef361859