Add CobaltStorageLoader.setObserver() to notify when storage changes.
The construction of the inner StorageFileObserver is complicated by the
fact that the FileObserver(String) constructor was deprecated in API 29
in favor of FileObserver(File). Therefore, there's a static create()
factory method to call one constructor or the other, and that led to the
inner class being static so it needs to keep its own filesDir File
object rather than being able to access the one in its containing
StorageFileObserver instance.

Note: The above paragraph doesn't fully apply to the cherry-pick since
lts.20 is still targeting API 28. The StorageFileObserver constructor
taking a File is not in the cherry-pick, but the static factory create()
method remains and it just calls the constructor with a String.

b/136026790

Change-Id: I03a75003f3a0d21cfff6cb3d3671309a0fd1533c
(cherry picked from commit 938637b14943f46ff8a3f1caa56bead6022e43f0)



Former-commit-id: d0658d3e1c42341cc9b718a0f4e32fbe2d3c33a5