Add CobaltStorageLoader.setObserver() to notify when storage changes.
The construction of the inner StorageFileObserver is complicated by the
fact that the FileObserver(String) constructor was deprecated in API 29
in favor of FileObserver(File). Therefore, there's a static create()
factory method to call one constructor or the other, and that led to the
inner class being static so it needs to keep its own filesDir File
object rather than being able to access the one in its containing
StorageFileObserver instance.

b/136026790

Change-Id: I03a75003f3a0d21cfff6cb3d3671309a0fd1533c

Former-commit-id: 5cfc58a75f2dc95d3c3a211db66081e6315ab966