Make IntersectionObserver emit an initial entry
When an intersection observer observes a new target, record a
mutation for the document so as to trigger a dirty layout. This will
allow layout::GenerateRenderTreeFromBoxTree() and Box::RenderAndAnimate()
to be asynchronously called, which is how new intersections are detected.

Make sure to also invalidate all the layout boxes associated with
new/removed roots/targets, so that added/removed intersection observer
objects are accordingly attached/detached from their associated box and
so that RenderAndAnimate() gets called on the box referencing the
intersection observer targets.

This cl also adds a unit test for this bug.

b/147912066

Change-Id: Ifef3f1e6ec733e0cdab47560d29975e50b82e3ba

Former-commit-id: 35f99e31b2e582d05a7447ba3824965533ff48bb