Create mechanism to optionally include tests in Python, GYP
This change implements the ability to optionally pull code into Python
and GYP depending on whether or not we are open-source. The current
implementation is meant to be used to include tests, but could be
expanded or used as a template for any similar purpose.

This change moves towards hiding all details about code that we are not
ready to release to open-source, until we are ready to make the push.
This fixes an immediate bug where elf_loader_test and
installation_manager_test are included depending on the existence of
certain directories. This causes inconsistencies between what we run and
what we build, which manifests as an error on mobile harness.

Please see the bug for more details.

b/148382535

Change-Id: I96a0c75509c3dfacf460fec6ec1b82e3037d7fb3

Former-commit-id: 63c8306787777a937a94435284d4c6ec474be99e