Silence some logs.
This cleans up a NOTIMPLEMENTED, and changes some logs to log once.

This should reduce black box test log sizes by ~9% and remove three
lines for each layout tests.

This also triggers a C4706 'assignment within conditional expression'
warning form the MSVS compiler from the assignment in the LOG_ONCE()
macro.  It's unclear why this warning isn't triggered for the other
uses of the same macro, but the compiler documentation suggests a
workaround that works and is included in this CL.

b/138843162

Change-Id: Idb410d5a23085fa0a9b13dc18c094878a2fb5f72

Former-commit-id: 65a543e41f019d179ed610e3790bef07da5f6f7d