Fix incorrect way to fetch ps4 netmask
sceNetCtlGetInfo is used to get the IP and netmask of the device.
However, it returns only either the IP or the netmask, depending
on its argument. The result is stored in a union (info) and so
info.ip_address and info.netmask always contain the same data.
The code was only calling sceNetCtlGetInfo once to get the IP
address, and then use info.netmask later on, which would still
contain the IP address and not the netmask.

Also, the unit tests check the highest bit to see if the netmask
starts with 0. It should therefore check the MSB with 0x80 and
not 0x8.

b/147029557

Change-Id: I833e8e08024787d2d000886ac06504998516e9b3

Former-commit-id: 205aed5b72c010da30501c068c90a4e1896828f7