Disable V8cHeapTracer during shutdown and reset after
This is the fix for the following bug. Kaido and I think the hanging
might be caused by V8 waiting for embedder to trace alive objects when
we already reset the heap tracer. This CL disables the tracer before
the shutdown GC so that V8 can clean up all Cobalt-created objects and
only reset the pointer after GC is complete.

b/148087942

Change-Id: I948498174b0da18161859e82920e8e60d742ecf9

Former-commit-id: 09ffa290c9b23af65303330112167fda1f03eede