[ps5] Fix build error on IsComplexIteratorForContainer
IsComplexIteratorForContainer() is used by EMUCaller and relies on that
std::iterator_traits<> has a member typedef named value_type, which is
not true on some platforms, include ps5.  It was used to branch to an
EMUCaller<> with less dependency on std::iterator_traits based on the
value of __cplusplus, now the EMUCaller<> with less dependency is
enabled on all Cobalt platform.

Note that EMUCaller is referenced during build but is not actually used
at runtime.

b/149065008

Change-Id: I9047984561341637aead583eb643306490e6070b

Former-commit-id: 488ce7cbb41cfa20bf56d2f7840ea893a2e67352