[Skia m79] Fix renderer_test crash on raspi
raspi still uses gcc 4.8, which has many issues with regards
to initialization syntax. When brackets instead of parentheses
are used in this particular situation, the copy constructor of
SkStrikeSpec gets wrongly called when initializing a const reference.

Added go/lbreview/150635062 as a followup to update the raspi compilers
and clean up this change.

b/149968865

Change-Id: Ibac34530facaf01ed2b25792c7a4071c71cce9f7

Former-commit-id: 13d6a2d5b056491b6e95095cbb4a7aa47fe5d20c