Add Skottie modules to Cobalt build
Create the .gypi files for the Skottie modules and any dependencies,
update the list of excluded files, make the list of included sources
more readable, and update the copyright headers.

This change also includes some minor backporting in RangeSelector to
make it C++11 compatible, and turns off usage of clang/gcc vector
extensions because this causes internal compiler errors on raspi.
(Added b/150947827 as a follow-up to to this issue.)

b/134191114

Change-Id: I86ba36d825c98c6df5ee02dedc83f25d2714a84c

Former-commit-id: e6bbcdf1ced66177663199ae757a2bb0313b21e9