Add error handling to SbAudioSinkPrivate
A new parameter is added to SbAudioSinkPrivate::Type::Create() with type
ErrorFunc.  The implementation can call this function when an error
occurred.  Currently there are only two types of errors supported:
fatal, or transient.  It is indicated by a boolean parameter to the
function.

This is implemented for Amazon's tunneled mode implementation on Android
TV, to report an audio sink capability change when the output
destination is changed.

Note that the error callback is only available in the internal
implementation used by AudioRendererSink, it hasn't been added to
SbAudioSink.

Change-Id: I8105235e3bb53b27bbfec2c7fc0bb25f27116776

Former-commit-id: 4ffe7a69cbaa5f96309205a3973423a3e2f69ba6