Strip evergreen-arm-hardfp deploy binaries.
This change establishes the pattern for creating a post-platform deploy
step to perform architecture specific actions. This implements binary
stripping for evergreen-arm-* platforms. In the future we should not
rely on the Raspberry Pi toolchain to accomplish this, but this fixes an
immediate problem. A TODO has been added to follow up.

b/151364508

Change-Id: I5d6c507ffc455033a9dbbda2a7948de5dca336ab

Former-commit-id: a31daa49def61224ba65d8a676c9df534186f0b1