Use absolute paths when setting Gradle properties
Don't rely on Java/Gradle to keep CWD the same as when it's invoked.

Prefix all Gradle params derived from GYP path variables with "$${PWD}/"
so that when ninja runs the deploy action, the CWD of the deploy target
will be prefixed to the relativized path of the GYP variable.

Go ahead and do the same for --cache and --project-dir for good measure,
even through it's not really necessary since those directly control
Gradle rather than setting properties for the build.

b/151379962

Change-Id: I0c7747574b31cf04062e3bc411654b08e3cc358f

Former-commit-id: 2fb1e0423456d550030ce73014b9eeaca5dac09f