Improve accuracy of audio time after seeking
During seeking, the target time can be in the middle of a DecodedAudio
object.  Now AudioRenderer will adjust the DecodedAudio object to make
it start exactly at the seek target time by removing any frames in the
beginning of the DecodedAudio object that are before the seek target
time.

Also fixed sample writing in SbMediaSetAudioWriteDurationTest.*, where
it used to start write from the second sample instead of the first one.

Change-Id: If5f417499d2412640b37598273a696eb3a48bcb2

Former-commit-id: c2e2ff1dd548ea7e50fce0f617acb55c3b9cff26