[ps5] Remove the workaround on max audio channel
SceAudioOutPortState.channels used to be always set to 8 even when there
isn't 5.1 output connected.  It was explicitly set back to 2 as a
workaround.

Now I was no longer able to reproduce this issue, with the new SDK on
non-alpha devkit.  The workaround is removed.

Also added logging for max audio output channels, and modified the
relevant nplb test to make it call SbAudioSinkGetMaxChannels() multiple
times.

b/147510458

Change-Id: I26aed2c1c222db1490af733a4d8167a83afc14b3

Former-commit-id: ead118c7702c22290f6256e481be8368c05c0283