Add no-op implementation of Overlay.SetShowViewportSizeOnResize command.
When collecting a trace in the Performance panel the frontend suspends
its OverlayModel and resumes it when the trace is stopped. Suspending
just disables the Overlay domain, but then when resuming after the
domain is re-enabled it tries to restore the previous settings and calls
Overlay.SetShowViewportSizeOnResize(). If that's not implemented it will
cause an uncaught exception in the frontend, which prevents analyzing
and displaying the collected trace data.

b/151643322

Change-Id: I4f568869d7d9f1f6ec3342d16d44c4a98d097595

Former-commit-id: 6830aef85a652c8d7b9985d61b3c7d823b82ca52