Starboardize, enable zip unittests.
This change starboardizes and enables the available zip unittests. In
addition, another test has been added that outputs the duration of time,
in milliseconds, that is taken to compress and decompress 32MiB in
memory.

The ZipTest.Zip and ZipTest.ZipIgnoreHidden are filtered for Android
because this platform does not correctly iterate through a directory
tree within the static content tree.

The ZipTest.Zip and ZipTest.ZipIgnoreHidden are filtered for PS4 because
this platform relies on SbDirectoryGetNext, which fails on buildbot.

b/145630516

Change-Id: I29d67b1cb447e3a518c8bb53f357eec8903dbd3d

Former-commit-id: 2e4771d2df98fb993c09b23030508a1cde8ac65f