Add content type to SbMediaIs{Audio|Video}Supported
So Starboard implementations can make decision on the full content
type.  For example, now the H5 player can detect whether the vp9 decoder
on the platform is software based or hardware based by using the
"features" attribute.

Note that we may make "features" a parameter to the above functions once
this is mature.

b/154356963

Change-Id: Ibb27875075e2b216eb7cb95e47c01b35c24fa56a

Former-commit-id: 950083a75ee0c2a082244b653f1eec9a9ccdd013