Update files for lottie-player demo & tests
This change creates a small demo that uses the lottie-player tag
(which can be updated as new features are implemented). This also
updates one of the files used in a pixel test, in order to ensure
that all of the samples that we use in Cobalt are open source.

b/155408365
b/134191114

Change-Id: I6d7969f3539da3d4af2d2e6de9480e53651e1a17

Former-commit-id: cdb707c37364ca46cd6c6591fdb0968855e7f0f7