[win32] Remove shell invocations from Gyp
This removes shell invocations in Gyp cross-platform codepaths,
with the intent of disabling shell calls entirely, in favor of
pymod_do_main.
The main driver is to be able to run Windows Gyp as a single
Python command invocation, rather than starting multiple
interpreters, and also remove dependency on Bash shell tools
presence on Win. This in turn has these benefits

- Containerization becomes viable, and does not require
  heavy custom tooling install into the container, i.e. just
  Win-native tools plus any Python version
- The Gyp invocations become overall much faster on Win and
  can be debugged under normal Python debug tools
- The Gyp code does become somewhat more testable and reusable
  as well
- Migration to another meta-build system like Cmake or GN
  becomes more straighforward as gyp_functions have 1:1
  equivalents

b/160261370

Change-Id: I0eb7e03d5cdb0689169563a356b2d203a4ef5d1d

Former-commit-id: 82409235a61b02fdd2cccfc9339d3cbab5e8fd93