Remove default max_cpu_in_bytes
With the deprecation of the GYP variable setting max_cobalt_cpu_usage
the auto_mem_setting variable max_cpu_in_bytes will never get set
with a setting of type kBuildType. The default value removed was
assumed to be the build setting, and so upon removal max_cpu_in_bytes
will (correctly) think it is not set, allowing it to use the
Starboard setting instead.

b/156484940

Change-Id: I1095a2da3826433222ed3b8e7e1f94a81cf9ab18

Former-commit-id: 6379d9efe47fb1f6108a67f7afe85766a2fdacd6