Revert media_set_audio_write_duration_test changes
This CL reverts a change where we decoupled the test code from using
internal-only code SbMediaIsAudioSupported(). By letting the test use
internal-only code, that means it cannot be used in nplb tests for
evergreen builds.

Reason for revert:

Due to an unforseen bug that is currently under investigation, the tests
fails on linux builders when using SbPlayerCreate as the main way of
testing whether a platform supports a certain media playback
configuration.

b/145746961
b/156382685

Change-Id: I995a01ceea7ff559eedd265e9437207334218709

Former-commit-id: 0841a743d249ca6b2509f517baffd7f092fbeaa5