Add Lottie "freeze" event
The Lottie web component triggers a "freeze" event and pauses
an animation when a playing animation goes offscreen. It triggers
a "play" event and resumes the animation from where it left off
when it becomes visible again.

This also applies when the element has "display: none", but not
when it has "visibility: hidden".

b/156115005

Change-Id: I054fc9765d022f54969842ffeee271c959d485ee

Former-commit-id: 8960934377c54f1f472a5575ef7b03a1f1ae4332