Fix egl.h and gles.h documentation on cobalt.dev.
This change uses a manual fix to clean up how Doxygen rendered the
function pointer members within structs in the egl.h and gles.h files.
Normal documentation approaches did not seem to work, and this is a
relatively simple unblocking fix until a permanent solution is foind.

This was done by:

1) Condense all function signatures so that they each fit onto their own
single line.
2) Replace *s with \\* to prevent them being understood as markdown
formatting.
3) Where A is the name of the function, and B is the entire signature,
updated each line to be "A; /**< B */".

b/112700312
b/125482921

Change-Id: I074f28c9c2f4c46618b11d9b8b19e0aadfbebc44

Former-commit-id: 7a8792d5dc073b11d8330df51280fe91a2142e2f