Inject TextToSpeechSettingsChanged event on Android
In Cobalt 21, we introduced the
kSbEventTypeAccessiblityTextToSpeechSettingsChanged event which is
required to be fired upon any system TTS settings change on
platforms supporting TTS. This change removed the obsolete logic to
relaunch the app when screen reader is turned on and replaced it with
code to fire the dedicated Starboard event for any TalkBack changes. See
bug below for why we can remove the relaunch logic.

b/158611551

Change-Id: I8ee8d0821150274c26f5a44a341c13b3bf6f3454

Former-commit-id: 1ebf95d428d8fa2626b52fd4603ce239fab37888