Add Lottie "hover" attribute
If the "hover" attribute is present, then the animation
should play when the mouse hovers over the element, and
stop when the mouse leaves the element.

This CL also includes a small change in SkottieAnimation
to make sure that stopped animations start from the beginning,
and not from where they were stopped, when they play again.

b/157245209

Change-Id: I3986bff49d5d9d52385f2121664fe01a9b8142be

Former-commit-id: 8e50101a158f465dffd197b48384284e995b681f