Freeze/Thaw TracingAgent to profile through a reload.
Save the string vector of categories that was set when tracing was
started so that they can be restored if tracing needs to be re-started
after navigation.

I had to deal with an issue this exposed with our hybrid JavaScript
agent architecture. DebugDispatcher disables stopping on exceptions
while running any internal hybrid JavaScript command, which requires the
V8InspectorSession that's not available until the ScriptDebuggerAgent
re-attaches the thawed state to the V8cScriptDebugger.

That led to a race when DevTools sends |Target.setAutoAttach| right
after navigation. We don't implement that domain, but DebugDispatcher
was trying to find the (unimplemented) JavaScript agent for it before
the V8InspectorSession was ready. To address that, DebugDispatcher now
doesn't even try to run commands for domains until the C++ agent has
registered itself.

b/160025996

Change-Id: I29c77192176566575add4da9cba19665f10201b1

Former-commit-id: 4f7879d6b31384d7ae03590184303fdb0707bb15