Convert ccache and starboard-toolchains to volumes
Currently we use the host's ccache and starboard-toolchains directories
as the mount points for the container's versions of these directories.
However, it is more portable if we are able to keep these as volumes.
This allows the volume driver to share it across multiple containers
more easily, when part of the container orchestration infrastructure.

b/161376210

Change-Id: I92feeaae564d096a4f7d0b3dc420b47abb2c4f74

Former-commit-id: fb6ad265a2730023910761c924bd66da9e7858e1