Fix misconfigured settings due to SB refactor
The GYP migration to the configuration API was done manually, and so
there were some settings that were overlooked when misconfigured. I
created a script comparing the settings before and after the migration
change in go/lbreview/154944 to look for discrepancies, and these were
the two that had changed, all others should be consistent with their
previous setting.

b/161565023

Change-Id: I597b5f2a00bdf1fc94e00de8f659ed9a68cdf240

Former-commit-id: 3ed194cc59afd77757c73eae2acc9a3fa0c4e3fd