Drain file timestamps are now in seconds, name length checked.
This change updates drain files to have their timestamps in seconds
instead of microseconds, and their prefix is now "d_". The length of the
filename that will be used is now checked against kSbFileMaxName.

This change also includes an additional Evergreen compliance test that
ensures we will have at least 64 bytes for a filename.

b/162115458

Change-Id: I5383a2571f741f6f9dc9bb3c50b251adaedd32af

Former-commit-id: e7448c500514ed708ad9a4c663929d612ecd6005