Windows base and win32 Docker containers.
This is a starting point for Windows build containers.
It contains all software neccessary to complete a Cobalt build.
It doesn't fully work yet however, as our Gyp config still has some
unix-specific commands for copy and touch.

The build.cmd file runs gyp which completes with a successful return
code, so this should work for test deployment on GCP Kubernetes.

Change-Id: I929b89a86fe0a1b051868338315061a1da8c5c10

Former-commit-id: e3dc944f5ada6a4c2b4f6cf7e247a9ffce7b7e9c